add error handling for failing documents during indexing
Added exception handling for failing documents during indexing in the `PopulateIndex` command. When a document fails to be indexed, an error message is displayed, and depending on the value of `should_skip_failing_documents` configuration option, indexing either continues or is aborted. If `should_skip_failing_documents` is set to `false`, a `DocumentFailedException` is thrown. Also added a new `DocumentFailedException` class to handle this type of exception.

Closes #54